fix(desktop): 输入框击键不再重渲染整个聊天界面 (#1021) (#1041)
ChatConsole 顶层的 input 状态让每次击键都重跑整个组件体并 reconcile
整棵聊天树；叠加 MessageBubble 的 memo 被 `?? []` 击穿与未 memo 的
sourcesSig，长会话（数百条消息）下打字/删除明显卡顿。

- 输入状态下沉到新组件 features/chat/Composer.tsx；ChatConsole 只保留
  onSubmit/streaming 等 props，经 ComposerHandle (setText/clear/focus)
  驱动全部 10 个程序化写入点（会话切换清空、发送成功清空、5 处草稿恢复、
  论文下载指示、编辑重答与重试预填）
- 发送文本复用既有 programmaticTextRef 传递，不再依赖 input state
- MessageBubble 的 sources 回退改用模块级 EMPTY_SOURCES 常量，memo 恢复生效
- sourcesSig 包 useMemo；已核实流式期间 messages 始终换新数组引用，
  来源提取刷新不受影响
- 顺带修复 tests/e2e/billing-hosted-live.spec.ts 的 prettier 漂移
  （develop 上 check-format 已因该文件失败）

实测（60 气泡会话，输入框敲 10 个字符 = 10 次 onChange）：
  MessageBubble 渲染次数增量  修复前 600（= 60 气泡 × 10）→ 修复后 0
